Questionable decisions: moved comment/vote count to mapper
This means we can't forget to add this data to the response object
(unless we're not using the mapper for some reason) but architecturally
I'm not sure this is an entirely sound decision.

Now the mapper isn't *purely* a mapper but I think I prefer this to
adding yet another class for processing posts etc.

Maybe.

We'll see if I revert this in an hour. :^)